Move addTask function to seperate JS file
