+++ src/add-remove.js
@@ -1,6 +1,12 @@
-// function addTask() {
-
-// }
+export default function addTask(e) {
+  if (e.key === 'Enter') {
+    const newTask = {};
+    newTask['description'] = this.value;
+    newTask['completed'] = false;
+    newTask['index'] = tasks.length + 1;
+    tasks.push(newTask);
+  }
+};
 
 // function removeTask() {
